Re-organized project (again)
recordStats/ and viewStats/ are gone now. Instead, we have:
stats/
    recording and retrieving log events happen here
statsPage/
    pages that show some stats now live here. These pages use stats/ REST API to get the information they show.
