+++ stats/admin.py
@@ -1,4 +1,4 @@
-from usageModel.models import *
+from models import *
 from django.contrib import admin
 
 admin.site.register(User)
