+++ statsPage/templates/showlog.html
@@ -28,7 +28,7 @@
                     "bJQueryUI": true,
                     "sPaginationType": "full_numbers",
                     "aaSorting": [[ 1, "desc" ]],
-                    "sAjaxSource": "json/country/?days=0",
+                    "sAjaxSource": "log/json/country/?days=0",
                     "sAjaxDataProp": "countries"
                 } );
             } );
@@ -38,7 +38,7 @@
                     "bJQueryUI": true,
                     "sPaginationType": "full_numbers",
                     "aaSorting": [[ 1, "desc" ]],
-                    "sAjaxSource": "json/domain/?days=0",
+                    "sAjaxSource": "log/json/domain/?days=0",
                     "sAjaxDataProp": "domains"
                 } );
             } );
@@ -48,7 +48,7 @@
                     "bJQueryUI": true,
                     "sPaginationType": "full_numbers",
                     "aaSorting": [[ 1, "desc" ]],
-                    "sAjaxSource": "json/platform/?days=0",
+                    "sAjaxSource": "log/json/platform/?days=0",
                     "sAjaxDataProp": "platforms"
                 } );
             } );
@@ -58,7 +58,7 @@
                     "bJQueryUI": true,
                     "sPaginationType": "full_numbers",
                     "aaSorting": [[ 2, "desc" ]],
-                    "sAjaxSource": "json/source/details/?days=0",
+                    "sAjaxSource": "log/json/source/details/?days=0",
                     "sAjaxDataProp": "detailedSources",
                     "bAutoWidth": false,
                     "aoColumns": [
@@ -77,7 +77,7 @@
                     "bJQueryUI": true,
                     "sPaginationType": "full_numbers",
                     "aaSorting": [[ 0, "desc" ]],
-                    "sAjaxSource": "json/eventlist/",
+                    "sAjaxSource": "log/json/eventlist/",
                     "sAjaxDataProp": "logs",
                     "aoColumns": [
                         { "mData": "date" },
