+++ statsPage/urls.py
@@ -1,0 +1,22 @@
+from django.conf.urls import patterns, include, url
+from django.views.generic import DetailView, ListView
+
+urlpatterns = patterns('statsPage.views',
+    # /log/
+    url(r'^$','show_log'),
+
+    # /log/errors
+    url(r'^errors/$','show_error_log'),
+
+    # /log/error/203215
+    url(r'^error/(?P<error_id>\d+)/$','show_error_details'),
+
+    # /log/login/
+    url(r'^login/$','show_sign_in_page'),
+
+    # /log/debug/
+    url(r'^debug/$', 'show_debug'),
+
+    # /log/debugerr/
+    url(r'^debugerr/$', 'show_debug_error'),
+)
