+++ statsPage/views.py
@@ -1,0 +1,126 @@
+import datetime
+import hashlib
+from pygeoip import GeoIP, GeoIPError
+from random import choice, randint
+import os
+import re
+import socket
+import sys
+import threading
+from django.conf import settings
+from django.contrib.auth import authenticate, login
+from django.core import serializers
+from django.core.exceptions import ObjectDoesNotExist
+from django.db import IntegrityError
+from django.db.models import Count
+from django.http import Http404, HttpResponse, HttpResponseRedirect
+from django.shortcuts import render_to_response, get_object_or_404
+from django.template import Context, loader, RequestContext
+from django.utils import simplejson, timezone
+from django.utils.datastructures import MultiValueDictKeyError
+from django.views.decorators.csrf import csrf_exempt
+from stats.models import *
+from django.conf import settings
+if not settings.configured:
+    settings.configure()
+
+
+
+####### TEMPLATE RENDERERS #######
+def show_sign_in_page(request):
+    '''
+    Shows login page.
+    '''
+    try:
+        username = request.POST['username']
+        password = request.POST['password']
+    except:
+        return render_to_response('authentication_page.html', {
+        }, context_instance = RequestContext(request))
+
+    # try logging in
+    user = authenticate(username = username, password = password)
+
+    # Invalid login
+    if user is None:
+        return render_to_response('authentication_page.html', {
+            'error_message': "Invalid username or password. Please try again.",
+        }, context_instance = RequestContext(request))
+    # De-activated user
+    elif not user.is_active:
+        return render_to_response('authentication_page.html', {
+            'error_message': "The account you are trying to use has been disabled.<br/>" + 
+            "Please contact a system administrator.",
+        }, context_instance = RequestContext(request))
+    # Valid login, active user
+    else:
+        login(request, user)
+        return HttpResponseRedirect('../')
+
+
+
+def show_log(request):
+    '''
+    Renders the logs.
+    '''
+    return render_to_response('showlog.html', {
+    }, context_instance = RequestContext(request))
+
+
+
+def show_error_log(request):
+    '''
+    Renders the logs.
+    '''
+    return render_to_response('showerrorlog.html', {
+    }, context_instance = RequestContext(request))
+
+
+
+def show_error_details(request, error_id):
+    '''
+    Renders the template which shows detailed error info.
+    '''
+    error_obj = get_object_or_404(Error, pk=error_id)
+    # countryLog = LogEvent.objects.filter(date__range = (date_from, timezone.now())).values('netInfo__country').annotate(count=Count('netInfo__country'))
+    actions = LogEvent.objects.values('date', 'action__name').filter(date__lte=error_obj.date, user=error_obj.logEvent.user, machine=error_obj.logEvent.machine).order_by('-date')[:10]
+    for action in actions:
+        action['name'] = action['action__name']
+        del(action['action__name'])
+        
+    return render_to_response('errordetails.html', {
+        "id": error_id,
+        "description": error_obj.description,
+        "stack_trace": error_obj.stackTrace,
+        "severity": error_obj.severity,
+        "user_comments": error_obj.userComments,
+        "execution_log": error_obj.executionLog,
+        "date": error_obj.date,
+        "source": error_obj.logEvent.source,
+        "platform": error_obj.logEvent.machine.getPlatform(),
+        "actions": actions,
+    }, context_instance = RequestContext(request))
+
+
+
+def show_debug(request):
+    '''
+    For debugging use only, will show a form where you can submit log events.
+    '''
+    if settings.DEBUG:
+        return render_to_response('debug.html', {
+        }, context_instance = RequestContext(request))
+    else:
+        raise Http404
+
+
+
+def show_debug_error(request):
+    '''
+    For debugging use only, will show a form where you can submit errors to be logged.
+    '''
+    if settings.DEBUG:
+        return render_to_response('debugerr.html', {
+        }, context_instance = RequestContext(request))
+    else:
+        raise Http404
