+++ usage/settings.py
@@ -80,9 +80,8 @@ INSTALLED_APPS = (
     'django.contrib.sites',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'recordStats',
-    'viewStats',
-    'usageModel',
+    'stats',
+    'statsPage',
     # Uncomment the next line to enable the admin:
     'django.contrib.admin',
     # Uncomment the next line to enable admin documentation:
