+++ usage/urls.py
@@ -6,10 +6,9 @@ from django.contrib import admin
 admin.autodiscover()
 
 urlpatterns = patterns('',
+    url(r'^$', include('statsPage.urls')),
+    url(r'log/', include('stats.urls')),
     (r'^media/(?P<path>.*)$','django.views.static.serve',{'document_root': settings.MEDIA_ROOT}),
-    # I'm still not sure if I like these split out as separate apps...
-    url(r'log/add/', include('recordStats.urls')),
-    url(r'^log/', include('viewStats.urls')),
 
     # Uncomment the admin/doc line below to enable admin documentation:
     # url(r'^admin/doc/', include('django.contrib.admindocs.urls')),
